backend/requirements.txt
fastapi
uvicorn
sqlmodel
pydantic
psycopg2-binary
python-dotenv